Version 1.4
Updated the SCCSUpdateVersion function to use "sccs prt -y" to get the latest version of the opened file. This fixes a bug in the script. Also, this enables the script to find whether the file is there in SCCS or not. If the file in not in SCCS, now you will see "SCCS-Not in SCCS" entry in the statusline.