Version 1.6
Added diff for any 2 SCCS versions of the opened file. If u diff with current version, it splits the screen vertically with the opened file to show the diff.If u diff with 2 versions that are different from opened file, it splits the screen horizontally and show vertically split diff in upper window. Added better way to get the SCCS version to show in status line. Added silent execute wherever possible.